Move status codes to constants index file
